Require --name and --version for makefile builds
