+++ trunk/cli/src/commands/build.rs
@@ -1,8 +1,10 @@
 use super::SubCommand;
 use crate::commands::pgx::build_pgx;
+// use crate::commands::makefile::build_makefile;
 use async_trait::async_trait;
 use clap::Args;
 use std::path::Path;
+use anyhow::anyhow;
 use tokio_task_manager::Task;
 use toml::Table;
 
@@ -12,6 +14,10 @@ pub struct BuildCommand {
     path: String,
     #[arg(short = 'o', long = "output-path", default_value = "./.trunk")]
     output_path: String,
+    #[arg(long = "version")]
+    version: Option<String>,
+    #[arg(long = "name")]
+    name: Option<String>,
 }
 
 #[async_trait]
@@ -25,9 +31,24 @@ impl SubCommand for BuildCommand {
             let dependencies = cargo_toml.get("dependencies").unwrap().as_table().unwrap();
             if dependencies.contains_key("pgx") {
                 println!("Detected that we are building a pgx extension");
+                if self.version.is_some() || self.name.is_some() {
+                    return Err(anyhow!("--version and --name are collected from Cargo.toml when building pgx extensions, please do not configure"));
+                }
                 build_pgx(path, &self.output_path, cargo_toml, task).await?;
                 return Ok(());
             }
+        }
+
+        // Check for Makefile
+        if path.join("Makefile").exists() {
+            println!("Detected a Makefile, guessing that we are building a C extension with 'make', 'make install...'");
+            // Check if version or name are missing
+            if self.version.is_none() || self.name.is_none() {
+                println!("Error: --version and --name are required when building a makefile based extension");
+                return Err(anyhow!("--version and --name are required when building a makefile based extension"));
+            }
+            // build_c_extension(path, &self.output_path, task).await?;
+            return Ok(());
         }
         println!("Did not understand what to build");
         Ok(())
