+++ trunk/cli/src/commands/mod.rs
@@ -4,7 +4,9 @@ use tokio_task_manager::Task;
 pub mod build;
 pub mod install;
 mod pgx;
+// mod makefile;
 pub mod publish;
+mod containers;
 
 #[async_trait]
 pub trait SubCommand {
