+++ trunk/cli/src/commands/pgx.rs
@@ -33,6 +33,7 @@ use tokio::task::JoinError;
 use tokio_stream::wrappers::ReceiverStream;
 use tokio_task_manager::Task;
 use toml::Value;
+use crate::commands::containers::ReclaimableContainer;
 
 #[derive(Error, Debug)]
 pub enum PgxBuildError {
@@ -96,43 +97,6 @@ fn semver_from_range(pgx_range: &str) -> Result<String, PgxBuildError> {
 
     let pgx_version = pgx_semver.to_string();
     Ok(pgx_version)
-}
-
-/// Used to stop container when dropped, relies on using [tokio_task_manager::TaskManager::wait]
-/// to ensure `Drop` will run to completion
-struct ReclaimableContainer<'a> {
-    id: &'a str,
-    docker: Docker,
-    task: Task,
-}
-
-impl<'a> ReclaimableContainer<'a> {
-    #[must_use]
-    pub fn new(name: &'a str, docker: &Docker, task: Task) -> Self {
-        Self {
-            id: name,
-            docker: docker.clone(),
-            task,
-        }
-    }
-}
-
-impl<'a> Drop for ReclaimableContainer<'a> {
-    fn drop(&mut self) {
-        let docker = self.docker.clone();
-        let id = self.id.to_string();
-        let handle = tokio::runtime::Handle::current();
-        let mut task = self.task.clone();
-        handle.spawn(async move {
-            println!("Stopping {id}");
-            docker
-                .stop_container(&id, None)
-                .await
-                .expect("error stopping container");
-            println!("Stopped {id}");
-            task.wait().await;
-        });
-    }
 }
 
 pub async fn build_pgx(
